[LayoutInstability] Cleanup test observe-layout-shift.html
Currently, the test contains two async_tests, which is not good because
they run in parallel so the tests can affect each other. The first test
is redundant with buffer-layout-shift.html. The second test uses a
helper function from EventTiming, so we import the script instead of
copy-paste and we rename the test.

Change-Id: Ibe6d200441a07c603bccf0858406cf2c4f1d37a4